refactored cancellation.ts and extracted out code that require fs to other file.
